requirements.txt
Flask==3.0.3
pymongo==4.10.1
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3
pdfplumber==0.11.4
PyPDF2==3.0.1
nltk==3.9.1
